server update

+++ DatabaseCursisten/DatabaseCursisten/Cursusprijzen.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Table2]
 (
-	[Prijs Ac ursus] INT NULL PRIMARY KEY, 
-    [Prijs B Cursus] INT NULL, 
-    [Prijs C Cursus] INT NULL
+	[Prijs Ac ursus] DECIMAL NULL PRIMARY KEY, 
+    [Prijs B Cursus] DECIMAL NULL, 
+    [Prijs C Cursus] DECIMAL NULL
 )
